+++ resources/js/Components/Pages/Shows/Elements/ShowIdIndexHeaderContents.vue
@@ -58,12 +58,12 @@
           <div
               class="w-full flex flex-wrap mt-5 m-auto xl:mx-0 items-center justify-center xl:justify-start gap-x-4 gap-y-2 ">
 
-            <ShowIdIndexPlayEpisode v-if="$page.props.user" :show="show" :team="team"/>
-            <ComingSoonShareAndSaveButtons/>
+            <ShowIdIndexPlayEpisode v-if="$page.props.auth.user" :show="show" :team="team"/>
+            <ComingSoonShareAndSaveButtons v-if="$page.props.auth.user" />
             <ShareButton :model="show" />
 
           </div>
-
+          <LoginToWatch/>
 
           <div class="description pt-10 w-full text-gray-300 text-center xl:text-left">
             <div class="w-full bg-gray-900 text-white tracking-wider text-sm">
@@ -90,6 +90,7 @@ import ShowIdIndexPlayEpisode from '@/Components/Pages/Shows/Elements/ShowIdInde
 import ShowsIdIndexWatchLive from '@/Components/Pages/Shows/Elements/ShowsIdIndexWatchLive.vue'
 import { onMounted } from 'vue'
 import ShareButton from '@/Components/Global/UserActions/ShareButton.vue'
+import LoginToWatch from '@/Components/Global/Banners/LoginToWatch.vue'
 
 const userStore = useUserStore()